下单出现数据格式错误

+++ src/main/java/com/sftc/web/service/impl/SFServiceAddressServiceImpl.java
@@ -152,7 +152,7 @@ public class SFServiceAddressServiceImpl implements SFServiceAddressService {
         Object weightObject = requestObject.get("weight");
         String weightStr = "1";
         if (weightObject != null) {
-            double weight = (Double) requestObject.get("weight");
+            int weight = (Integer) requestObject.get("weight");
             weightStr = (weight == 0 ? 1 : weight) + "";
         }
